+++ programs/pluto/state.c
@@ -857,7 +857,7 @@ void fmt_state(struct state *st, time_t n
 	if(st->hidden_variables.st_dpd) {
 	    time_t n = time(NULL);
 	    snprintf(dpdbuf, sizeof(dpdbuf), "; lastdpd=%lds(seq in:%u out:%u)"
-		     , st->st_last_dpd !=0 ? n - st->st_last_dpd : -1
+		     , st->st_last_dpd !=0 ? n - st->st_last_dpd : (long)-1
 		     , st->st_dpd_seqno
 		     , st->st_dpd_expectseqno);
 	} else {
