+++ programs/pluto/x509.c
@@ -745,7 +745,8 @@ check_validity(const x509cert_t *cert, time_t *until)
     if (current_time > cert->notAfter) {
 	char tbuf[TIMETOA_BUF];
 
-	DBG_log("  aftercheck : %ld > %ld", current_time, cert->notAfter);
+	DBG_log("  aftercheck : %ld > %ld", (unsigned long)current_time
+		, (unsigned long)cert->notAfter);
 	return builddiag("X.509 certificate expired at %s (it is now %s)"
 			 , timetoa(&cert->notAfter, TRUE, tbuf, sizeof(tbuf))
 			 , curtime);
